+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 - Add better `panic!` message to `map_range` if cast fails.
 - Add many items to prelude (audio, io, math, osc, ui, window).
 - Change event positioning types to use DefaultScalar.
+- Implement `draw.polygon()`
+- Update internal `IntoDrawn` API to support a dynamic number of arbitrary
+  vertices.
+- Update `Drawing` API to allow builders to produce new `Drawing` types.
 
 # Version 0.6.0 (2017-06-07)